requirements/requirements.txt
-i https://pypi.org/simple
-e .
aioredis==1.2.0
amqp==2.4.2
asgiref==2.3.2
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.3.3
automat==0.7.0
backcall==0.1.0
billiard==3.5.0.5
birdy==0.3.2
boto3==1.9.130
botocore==1.12.130
celery==4.2.0
certifi==2019.3.9
channels-redis==2.3.1
channels==2.1.5
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
csscompressor==0.9.5
daphne==2.2.5
decorator==4.4.0
django-appconf==1.0.3
django-celery-results==1.0.1
django-ckeditor==5.6.1
django-compressor==2.2
django-cors-headers==2.5.2
django-environ==0.4.4
django-extensions==2.1.3
django-filter==2.1.0
django-js-asset==1.2.2
django-mptt==0.9.1
django-paypal==1.0.0
django-redis==4.10.0
django-rest-swagger==2.2.0
django-storages==1.7.1
django-widget-tweaks==1.4.3
django==2.2
djangorestframework-simplejwt==4.1.3
djangorestframework==3.9.2
docutils==0.14
hiredis==1.0.0
httplib2==0.12.1
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
ipython-genutils==0.2.0
ipython==7.0.1
itypes==1.1.0
jedi==0.13.3
jinja2==2.10
jmespath==0.9.4
kombu==4.5.0
mailchimp3==3.0.7
markupsafe==1.1.1
msgpack==0.5.6
oauth2==1.9.0.post1
oauthlib==3.0.1
openapi-codec==1.3.2
parse==1.8.4
parso==0.4.0
pexpect==4.6.0 ; sys_platform != 'win32'
pickleshare==0.7.5
pillow==5.1.0
prompt-toolkit==2.0.9
psycopg2-binary==2.7.4
ptyprocess==0.6.0
pydash==4.7.4
pygments==2.3.1
pyhamcrest==1.9.0
pyjwt==1.7.1
python-dateutil==2.8.0 ; python_version >= '2.7'
pytz==2018.9
rcssmin==1.0.6
redis==3.2.1
requests-oauthlib==1.2.0
requests==2.20.0
rjsmin==1.0.12
s3transfer==0.2.0
sentry-sdk==0.7.10
simplegeneric==0.8.1
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.2
twisted==18.9.0
txaio==18.8.1
uritemplate==3.0.0
urllib3==1.24.1 ; python_version >= '3.4'
uwsgi==2.0.18
vine==1.3.0
wcwidth==0.1.7
zope.interface==4.6.0

# dev packages
-i https://pypi.org/simple
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
coverage==4.5.1
decorator==4.4.0
django-debug-toolbar==1.10.1
django==2.2
factory-boy==2.11.1
faker==1.0.4
flake8==3.5.0
ipython-genutils==0.2.0
ipython==7.0.1
isort==4.3.4
jedi==0.13.3
mccabe==0.6.1
more-itertools==7.0.0
parso==0.4.0
pexpect==4.6.0 ; sys_platform != 'win32'
pickleshare==0.7.5
pluggy==0.9.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
py==1.8.0
pycodestyle==2.3.1
pyflakes==1.6.0
pygments==2.3.1
pytest-asyncio==0.9.0
pytest-cov==2.6.0
pytest-django==3.4.3
pytest==3.9.2
python-dateutil==2.8.0 ; python_version >= '2.7'
pytz==2018.9
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
traitlets==4.3.2
wcwidth==0.1.7
